requirements.txt
Flask==2.3.2
flask-restx==1.1.0
openai==1.14.3
redis==5.0.4
python-dotenv==1.0.1
werkzeug==2.3.7
